% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householder_acoustic_analysis.R
\name{Householder}
\alias{Householder}
\title{Householder Acoustic Analysis}
\usage{
Householder(A, x)
}
\arguments{
\item{A}{Matrice carrée représentant le système acoustique}

\item{x}{Vecteur représentant le signal}
}
\value{
Une liste contenant :
Q : matrice orthogonale,
R : matrice triangulaire supérieure,
H : matrice de Householder,
Ax : produit original,
QRx : produit reconstruite via QR
}
\description{
Calcule la matrice de Householder, la décomposition QR, et analyse un signal acoustique.
}
