% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{householder_acoustic_shiny}
\alias{householder_acoustic_shiny}
\title{Interface Shiny pour HouseholderAcoustics}
\usage{
householder_acoustic_shiny()
}
\description{
Lance une interface interactive pour tester le package sans écrire de code.
}
