% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fft.R
\name{internal_fft_analysis}
\alias{internal_fft_analysis}
\title{Analyse fréquentielle d’un signal acoustique}
\usage{
internal_fft_analysis(signal, fs = 1)
}
\arguments{
\item{signal}{Vecteur numérique représentant un signal temporel.}

\item{fs}{Fréquence d’échantillonnage (Hz).}
}
\value{
Une liste contenant les fréquences et le spectre d’amplitude.
}
\description{
Calcule le spectre fréquentiel d’un signal acoustique
par transformée de Fourier rapide (FFT).
}
\details{
L’analyse fréquentielle est essentielle pour l’identification
du bruit, des résonances et des fréquences dominantes.
}
\keyword{FFT}
\keyword{acoustics}
\keyword{signal-processing}
