% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_report.R
\name{internal_generate_report}
\alias{internal_generate_report}
\title{Génération automatique d’un rapport acoustique}
\usage{
internal_generate_report(results)
}
\arguments{
\item{results}{Liste des résultats issus de l’analyse.}
}
\value{
Fichier PDF généré.
}
\description{
Génère un rapport scientifique PDF contenant
les calculs, graphiques et interprétations acoustiques.
}
\details{
Le rapport est adapté pour un usage académique :
projet, mémoire ou publication scientifique.
}
\keyword{acoustics}
\keyword{report}
