% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_householder.R
\name{internal_householder_matrix}
\alias{internal_householder_matrix}
\title{Construction de la matrice de Householder}
\usage{
internal_householder_matrix(v)
}
\arguments{
\item{v}{Vecteur de Householder.}
}
\value{
Matrice orthogonale de Householder.
}
\description{
Génère la matrice de réflexion orthogonale H = I - 2vvᵀ.
}
\details{
Cette matrice est utilisée pour les projections orthogonales
et l’orthogonalisation des signaux acoustiques.
}
\keyword{internal}
