% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_householder.R
\name{internal_householder_vector}
\alias{internal_householder_vector}
\title{Construction du vecteur de Householder}
\usage{
internal_householder_vector(x)
}
\arguments{
\item{x}{Vecteur numérique.}
}
\value{
Vecteur normalisé définissant la réflexion.
}
\description{
Calcule le vecteur de Householder définissant une réflexion
orthogonale dans l’espace des signaux acoustiques.
}
\details{
En acoustique, ce vecteur permet d’annuler ou de réorienter
les composantes énergétiques dominantes d’un signal.
}
\keyword{internal}
