% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_qr.R
\name{internal_qr_householder}
\alias{internal_qr_householder}
\title{Décomposition QR par Householder}
\usage{
internal_qr_householder(A)
}
\arguments{
\item{A}{Matrice numérique représentant un système acoustique.}
}
\value{
Une liste contenant :
\itemize{
  \item Q : matrice orthogonale (modes acoustiques)
  \item R : matrice triangulaire supérieure
}
}
\description{
Effectue la décomposition QR d’une matrice acoustique.
}
\details{
La décomposition QR permet l’analyse modale et la stabilité
numérique des systèmes acoustiques.
}
\keyword{QR}
\keyword{acoustics}
\keyword{internal}
