% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_visualization.R
\name{internal_visualization}
\alias{internal_visualization}
\title{Visualisation automatique des résultats acoustiques}
\usage{
internal_visualization(x, y, A, H, Q, R, Ax, QRx)
}
\arguments{
\item{x}{Signal acoustique original.}

\item{y}{Signal transformé.}

\item{A}{Matrice acoustique.}

\item{H}{Matrice de Householder.}

\item{Q}{Matrice orthogonale.}

\item{R}{Matrice triangulaire.}

\item{Ax}{Réponse directe.}

\item{QRx}{Réponse modale.}
}
\value{
Aucun (affichage graphique).
}
\description{
Génère l’ensemble des graphiques liés à l’analyse :
signaux temporels, matrices, réponses Ax et QRx.
}
\keyword{acoustics}
\keyword{visualization}
